+++ src/components/layout/home/WritePetition/index.tsx
@@ -5,8 +5,7 @@ import { WriteFeature } from 'features/home/write.feature';
 import Confirm from 'components/common/Dialog';
 
 import { useNavigate } from 'react-router-dom';
-import { useDialog } from 'hooks/useDialog';
-import { useModal } from 'hooks/useModal';
+import { useDialog, useModal } from 'hooks/useDialog';
 import * as S from './style';
 import CheckPetitionModal from './CheckPetitionModal';
