+++ src/hooks/useDialog.ts
@@ -1,4 +1,4 @@
-import { dialogState } from 'atoms/dialog.atom';
+import { dialogState, modalState } from 'atoms/dialog.atom';
 import { useRecoilState } from 'recoil';
 
 export const useDialog = () => {
@@ -11,4 +11,16 @@ export const useDialog = () => {
   };
 
   return { openDialog, closeDialog, dialog };
+};
+
+export const useModal = () => {
+  const [modal, setModal] = useRecoilState(modalState);
+  const openModal = () => {
+    setModal(true);
+  };
+  const closeModal = () => {
+    setModal(false);
+  };
+
+  return { openModal, closeModal, modal };
 };
